Add progress bars, refactor terms, write schema
The schema is the biggest change. I have been thinking about the schema
in an OOP fashion, which was quite wrong-headed. The domain, range, and
subClassOf terms are better suited for inference than data
specification. They can be used to infer type. For example, a pair of
terms related by the "has_segment" property are defined to be,
respectively, a Strain and Segment. The schema should not be an
enumeration of all the slots each class has, but rather is a succinct
statement of a few powerful relations from which all other types in the
system are inferred.