Rebuild with Token/Phrase/Interpreter
A Token object knows how to parse a exact match string and knows how it
is related to other tokens.

A Phrase is a list of Token objects that are semantically related.

An Interpreter looks at the structure of the document to infer types
(for example from the columns in a table) and then collates Token
objects into a list of Phrase objects.

Currently I have special Interpreter objects for parsing homogenous
lists and tables. These both use context across the column/list to
decide which Token matches the data (i.e. resolve ambiguities). I also
have a Datum Interpreter that interprets a string without context, a
heterogenous list interpreter to treats each field as a Datum, and a
ragged interpreter that takes a list of lists and treats each list as a
Phrase (e.g., a Fasta file with a variable number of fields in the
header).