Add constellation recipe
Sort of - actually this is a bit of a hybrid between a `mk_*` command
and what a recipe will be. The `const` subcommand generates a table of
constellations from the data currently in the database. I expect to use
it mostly to generate secondary data that I can feed back into the
database (e.g., add constellation annotations).