Make stub tag/get implementation
The basic idea here is to tag identifiers for retrieval in an update
operation and then retrieve then in a query. The tags are persist until
cleared. The `get-tagged*.rq` files are default queries. The user may of
course make their own by filtering to the `?xxx f:query_tag ?id`. Where
`?xxx` is not generally used.