Make subtype properly reuse prior subtypes
There is an ontological problem in the database. Both strain and epiflu
isolate can be be the subject of the `has_segment` relation. But one
strain can be related to multiple epiflu isolates. So for all the epiflu
strains, their strain level data is split between the isolate and strain
nodes.

This is how things are currently wired:
```
             {1,1-8}
<strain> -------------> <segment>
    |     has_segment
    |
    |        {1,1}
    '-----------------> "A/...."
        strain_name

                  {1,1-8}
<epi_isolate> -------------> <segment>
    |          has_segment
    |
    |   {n,1}
    '-----------------> "A/...."
        strain_name
```

The semantically logical wiring would instead be:
```
             {1,1}
<strain> -------------> "A/..."
    ||    strain_name
    ||
    |'-----> strain properties, like host
    |
    |      {1,n}                       {1,1-8}
    '------------------> <isolate> ------------> <segment>
                             |
                             |
                             '---> isolate properties like deposit date
```

The problem is that not all strains have epiflu ids, and in these cases
the isolate node would be an empty placeholder that would complicate
queries. But maybe this is the right thing to do.