Begin scripting builds
Ideally building and updating octofludb should be as simple as running
`octofludb update`. To this end, I am creating a YAML-based config
system where the database configuration and the pathways to directories
containing data can be stored. I am also removing all raw data that was
stored in this repository.

The analysis that is currently performed by build.sh will now be written
into the `octofludb update` script.