pull is fully implemented
Next thing to do is to thoroughly debug it.

To this end, I need to write a good debugging system for pulling
different crossections of the data and checking for inconsistencies or
missing slices. Of course, I also need to carefully unit test whatever I
can. This will involve looking for material that can be pulled into
pure, testable, units.

Then I need to specify a small and clean subset of the ontology and
ensure it follow cleanly and naturally from the input data. The main
issue here is the disconnect between isolates and strains.