Make levels optional
The `levels` field stores the tokens that may be related to one another.
If the `levels` field is set to None, then there is no restriction on
fields. I had removed the Optional status of the field earlier, but now
it is added back in and everything is dandy once again.

TODO:

 * figure out why lineage is turning up as a NaN double in the gisaid
   golden test.

 * add full genbank golden test

 * completely delete the octofludb home and all related files, then
   rebuild it all to ensure the latest versions are working.

 * create a golden test for each subcommand

 * create a v1.0.0 release